Make test suite importable
